Occlusion through Unity Audio Spatialization
Unity's built-in audio system allows us to create spatial audio without the need for scripts, which includes more realistic panning and a basic occlusion.

Changes:
-Disabled Audio 1D on NPC Prefab
-Enabled Spatialization on NPC Prefab
-Set NPC Audio Source SpatialBlend to 3D and Volume to 1
-Set NPC Audio Source rolloff to a 1-10 linear rolloff

The rolloff controls the distance at which panning takes place. Should absolutely be different, this is just a good number for making sure it works.