Timers and End Screens
Updated the timer scripts including some of the end screen and scoring scripts since the time functionality was not working. Also, re-added the multiple ways to get to the end screen (time runs out, 10 errors, or end of script is reached). For some reason, all time is moving twice as fast as it should, but I have taken that into account when tracking and displaying time (by doubling the level and wait times and by halving the remaining time on the display screen). Also, tweaked some timing on the conversation script, but the news audio is not playing, so that will need to be looked into.