New Score Algorithm
Scores are now set as so:
Score = 100 - (errors * 8) - ((time-par)/600), where time is how long the player took, and par is the fastest they could have beaten the level.

Due to a bug somewhere in LevelConstructor, level progress ins't being saved correctly, so DisplayScore gets confused on what value to give the Stars. This change can't be pushed until that bug is fixed.

Also, reduced engine noise by 50% for coasting and braking(?), and 33% for accelerating.