Fixed the 0 Progress Bug
The underlying issue (progress being set to 0 becuase currentDialogueStartTime is set as 0 in the final dialogue section) is still there, but the symptoms have been treated. CLFM will just manually set progress to 100 on level end.

Bug can still occur if the player fails the level in the final section.