Special Failure Dialogue
Dialogue will play when the player fails the level before transitioning to the fail screen.

Currently only plays a specific clip for Rosales, functionality for chosing the correct fail based on level and context has not been implemented.

May behave incorrectly if the player finishes the level while the fail dialogue is playing or manages to get 11 errors while the fail dialogue is playing.